Added all files for Data Project 1
Link for JSON is in the txt file provided, while the CSV file is the included US Census file. Project runs with user input, so for running files make sure the file is in the working directory before typing in the name of the file with extension. Collaborative effort between Matt Gottfried and Nick Moir.